Juegos2 completado

+++ Erronka/app/Http/Controllers/UsuariosController.php
@@ -4,14 +4,14 @@ namespace App\Http\Controllers;
 use App\Models\Usuarios;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use App\Models\Partidas;
 
 class UsuariosController extends Controller
 {
     //VER
     public function index()
     {
-        // $miUsu = Usuarios::where('usuario','=',session()->get('usuario'))->get();
-        $miUsu = Usuarios::where('id','=',1)->get();
+        $miUsu = session()->get('usuario');
         $usu = Usuarios::all();
         return view('Comercio.admin', compact('usu', 'miUsu'));
     }
@@ -30,34 +30,41 @@ class UsuariosController extends Controller
     //CREAR
     public function store(Request $request)
     {
-        // $request->validate([
-        //     'modelo' => 'required|max:40',
-        //     'nacionalidad' => 'required|max:40',
-        //     'anyo' => 'required|integer',
-        // ]);
+        if(Usuarios::where('usuario', $request->usuario)->exists()){
+            return redirect()->action([UsuariosController::class, 'create'])->with('error', 'Erabiltzailea existitzen da.');
+        }
+        
+        $request->validate([
+            'nombre' => 'required|between:3,50',
+            'apellido' => 'required|between:3,50',
+            'usuario' => 'required|between:6,50', // unique:posts|
+            'pass' => 'required|between:8,50',
+            'passB' => 'required|between:8,50',
+        ]);
 
         $usu = new Usuarios();
-        $usu->nombre = $request->nombre;
-        $usu->apellido = $request->apellido;
-        $usu->usuario = $request->usuario;
-        $usu->pass = $request->pass;
-        $usu->rol = 0;
-        $usu->foto = "default.png";
-        $usu->type = "png";
 
+        if ($request->pass == $request->passB) {
 
-        // if ($request->hasFile("img")){
-        //     $img = $request->file("img");
-        //     $nomImg = Str::slug($request->usuario).".".$img->guessExtension();
-        //     $ruta = public_path("img/profile/");
-
-        //     copy($img->getRealPath(),$ruta.$nomImg);
-
-        //     $usu->foto = $nomImg;
-        //     $usu->type = $img->guessExtension();
-        // }
-        $usu->save();
-        return redirect()->action([UsuariosController::class, 'create']);
+            $usu->nombre = $request->nombre;
+            $usu->apellido = $request->apellido;
+            $usu->usuario = $request->usuario;
+            $usu->pass = $request->pass;
+            $usu->rol = 0;
+            $usu->foto = "default.png";
+            $usu->type = "png";
+            $usu->save();
+    
+            $partidas = new Partidas();
+            $partidas->juego1 = 0;
+            $partidas->juego2 = 0;
+            $partidas->usuario = Usuarios::where('usuario','=', $request->usuario)->get()[0]->id;
+            $partidas->save();
+            
+            return redirect()->action([UsuariosController::class, 'create'])->with('bien', 'Erabiltzailea sortu egin da.');
+        } else {
+            return redirect()->action([UsuariosController::class, 'create'])->with('error', 'Pasahitza ez da berdina.');
+        }
     }
     //LOGIN
     public function logSes(Request $request){
@@ -70,15 +77,22 @@ class UsuariosController extends Controller
         //     }   
         // }
 
+        
+        
+        // $usuarios = Usuarios::all();
         $usuarios = Usuarios::where('usuario','=',$request->usuario)->get();
+        
         foreach($usuarios as $usu){
-            if($request->usuario != $usu->usuario && $request->pass != $usu->pass){
-                return view('Comercio.log-reg', 'El usuario es incorrecto');
+            if($request->pass == $usu->pass){
+                session(['usuario' => $usu]);
+
+                // Route::post('/juegos',  [PartidasController::class, 'store'])->name('Comercio.parCrear');
+                return view('Comercio.index');
             }   
         }
 
-        session(['usuario' => $usu]);
-        return view('Comercio.index');
+        $error = 'El usuario: ' . $request->usuario . ', es incorrecto o la contraseña no coincide';
+        return view('Comercio.log-reg', compact('error'));
     }
 
      //CERRAR SESION
@@ -88,19 +102,41 @@ class UsuariosController extends Controller
         return view('Comercio.index');
     }
     
+    // MEJORAR ROL
+    public function rol(Request $request, $id)
+    {
+        $codigoAdmin = 'ADMIN123';
+
+        if ($request->rol == $codigoAdmin) {
+            $usu = Usuarios::findOrFail($id);
+            $usu->rol = 1;
+            $usu->save();
+
+            session(['usuario' => $usu]);
+        }
+
+        return redirect()->action([UsuariosController::class, 'index']);
+    }
     
     //UPDATE
     public function update(Request $request, $id)
     {
+        $request->validate([
+            'nombre' => 'required|max:50',
+            'apellido' => 'required|max:50',
+            'usuario' => 'required|max:50', // unique:posts|
+            'pass' => 'required|max:50',
+        ]);
+        
         $usu = Usuarios::findOrFail($id);
         $usu->nombre = $request->nombre;
         $usu->apellido = $request->apellido;
         $usu->usuario = $request->usuario;
-        $usu->rol = $request->rol;
 
         if ($request->hasFile("img")){
-            unlink(public_path('img/profile/'.$usu->foto));
+            if($usu->foto != 'default.png') unlink(public_path('img/profile/'.$usu->foto));
             $img = $request->file("img");
+            
             $nomImg = Str::slug($request->usuario).".".$img->guessExtension();
             $ruta = public_path("img/profile/");
 
@@ -108,19 +144,11 @@ class UsuariosController extends Controller
 
             $usu->foto = $nomImg;
             $usu->type = $img->guessExtension();
-        } elseif ($request->usuario == $usu->usuario) {
-            
-            $img = public_path('img/profile/'.$usu->foto);
-            
-            $nomImg = Str::slug($request->usuario).".".$usu->type;
-            $ruta = public_path("img/profile/");
-
-            copy($img,$ruta.$nomImg);
-            unlink(public_path('img/profile/'.$usu->foto));
-            $usu->foto = $nomImg;
-        };
+        } 
 
         $usu->save();
+
+        session(['usuario' => $usu]);
         return redirect()->action([UsuariosController::class, 'index']);
     }
 
@@ -128,8 +156,12 @@ class UsuariosController extends Controller
     public function destroy(Request $request)
     {
         $usu = Usuarios::findOrFail($request->usuario);
-        unlink(public_path('img/profile/'.$usu->foto));
+        $part = Partidas::where('usuario',$request->usuario)->get()[0];
+
+        if($usu->foto != 'default.png') unlink(public_path('img/profile/'.$usu->foto));
+        $part->delete();
         $usu->delete();
-        return redirect()->route('Comercio.index')->with('success', 'Erabiltzailea ezabatu da.');
+        
+        return redirect()->route('Comercio.admin')->with('mensaje', 'Erabiltzailea ezabatu da.');
     }
 }
